Refactor:molecules, organisms 재분류
