+++ app/feature/curation/components/CurationDetail/organisms/CurationDetailInfoCard.tsx
