+++ app/feature/curation/components/CurationDetail/organisms/NoSavedPlace.tsx
