+++ app/feature/curation/components/CurationInfo/molecules/NoCurationText.tsx
