+++ app/feature/curation/components/CurationInfo/organisms/CurationInfoCardLight.tsx
