+++ app/feature/curation/components/CurationInfo/stories/CurationInfoCardDark.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import CurationInfoCardDark from "../molecules/CurationInfoCardDark";
+import CurationInfoCardDark from "../organisms/CurationInfoCardDark";
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
