+++ app/feature/curation/components/CurationInfo/stories/CurationInfoCardLight.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import CurationInfoCardLight from "../molecules/CurationInfoCardLight";
+import CurationInfoCardLight from "../organisms/CurationInfoCardLight";
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
