+++ app/feature/place/components/PlaceDetail/organisms/RelatedCurationSlider.tsx
