+++ app/feature/place/components/PlaceDetail/organisms/RelatedPlaceSlider.tsx
