+++ app/feature/place/components/PlaceInfo/organisms/PlaceInfoCardTop.tsx
@@ -15,7 +15,7 @@ import { myCurationModalInfoSelector } from "@common/state/myCurationModal";
 import Chip from "@common/components/ui/buttons/Chip/Chip";
 import ImageWrapper from "@common/components/ui/wrapper/ImageWrapper";
 import { validateLoggedIn } from "@common/utils/validate/validateLoggedIn";
-import PlaceInfoCardTopScrapIcon from "../atoms/PlaceInfoCardTopScrapIcon";
+import PlaceInfoCardTopScrapIcon from "../molecules/PlaceInfoCardTopScrapIcon";
 import CheckSmallIcon from "@common/assets/icons/check/CheckSmallIcon";
 import LineIcon from "@common/assets/icons/line/LineIcon";
