+++ app/feature/search/components/SearchResult/template/SearchResult.tsx
@@ -4,9 +4,9 @@ import Tab from "@common/components/ui/tab/Tab";
 import { lazy } from "react";
 import { SearchCurationResponse } from "@feature/search/queries/dto/search-type";
 import Divider from "@common/components/ui/divider/Divider";
-import CurationInfoCardLight from "@feature/curation/components/CurationInfo/molecules/CurationInfoCardLight";
+import CurationInfoCardLight from "@feature/curation/components/CurationInfo/organisms/CurationInfoCardLight";
 import useTextSearchBar from "../../../hooks/SearchText/useTextSearchBar";
-import SearchNoResult from "../molecules/SearchNoResult";
+import SearchNoResult from "../organisms/SearchNoResult";
 import dynamic from "next/dynamic";
 import useGetTextSearchPlaceData from "@feature/search/queries/useGetTextSearchPlaceData";
 import { useRecoilValue } from "recoil";
